Move some things around, fix typos
