+++ src/components/Section/ProposalVoting/ChangeVoteButton.tsx
@@ -4,7 +4,7 @@ import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 import { Link } from 'decentraland-gatsby/dist/plugins/intl'
 import { Button } from 'decentraland-ui/dist/components/Button/Button'
 
-import { Vote } from '../../entities/Votes/types'
+import { Vote } from '../../../entities/Votes/types'
 
 interface ChangeVoteButtonProps {
   vote: Vote | null
