+++ src/components/Section/ProposalVoting/DelegationsLabel.css
