+++ src/components/Section/ProposalVoting/DelegationsLabel.tsx
@@ -3,9 +3,12 @@ import React from 'react'
 import Markdown from 'decentraland-gatsby/dist/components/Text/Markdown'
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 
-import { DelegationsLabelProps } from '../../entities/Votes/utils'
-
 import './DelegationsLabel.css'
+
+export interface DelegationsLabelProps {
+  delegateLabel?: { id: string; values?: any }
+  delegatorsLabel?: { id: string; values?: any }
+}
 
 const DelegationsLabel = ({ delegateLabel, delegatorsLabel }: DelegationsLabelProps) => {
   const t = useFormatMessage()
