+++ src/components/Section/ProposalVoting/VotedChoiceButton.tsx
@@ -2,9 +2,15 @@ import React from 'react'
 
 import useFormatMessage from 'decentraland-gatsby/dist/hooks/useFormatMessage'
 
-import { VotedChoice } from '../../entities/Votes/utils'
-
 import ChoiceButton from './ChoiceButton'
+
+export interface VotedChoice {
+  id?: string
+  values?: any
+  delegate?: string
+  voteCount?: number
+  totalVotes?: number
+}
 
 const VotedChoiceButton = ({ id, values, delegate, totalVotes, voteCount }: VotedChoice) => {
   const t = useFormatMessage()
