+++ src/components/Section/ProposalVoting/VotingSectionFooter.css
