+++ src/entities/Votes/types.ts
