+++ src/intl/en.json
@@ -51,8 +51,8 @@
       "description_too_large": "Description field is too large",
       "choices_insufficient": "Insufficient options",
       "choices_empty": "Some option is empty",
-      "choices_too_short": "Some option is to short",
-      "choices_too_long": "Some option is to long",
+      "choices_too_short": "Some option is too short",
+      "choices_too_long": "Some option is too long",
       "submission_vp_not_met": "This action requires at least 100 VP."
     },
     "draft": {
