+++ src/modules/contracts/abi/MiniMe.abi.json
